backend/requirements.txt

Django==4.2.0
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
django-cors-headers==4.3.1
redis==5.0.1
Pillow==10.1.0
daphne
